kernel: replace unsafe uses of 1L
On some 64bit systems (most notably, Windows), sizeof(long) == 4, and so
expressions like `1L << 60` overflow. Moreover, an enum strictly
speaking is only guaranteed to be able to represent an int, and so using
it to represent 64bit constants is not safe; we hence turn
INT_INTOBJ_MIN and INT_INTOBJ_MAX into global static constants.

+++ src/integer.c
@@ -1524,7 +1524,7 @@ static Obj ProdIntObj ( Obj n, Obj op )
   /* <res> = 0 means that <res> is the neutral element                     */
   else if ( IS_INTOBJ(n) && INT_INTOBJ(n) >   1 ) {
     res = 0;
-    k = 1L << NR_SMALL_INT_BITS;
+    k = (Int)1 << NR_SMALL_INT_BITS;
     l = INT_INTOBJ(n);
     while ( 0 < k ) {
       res = (res == 0 ? res : SUM( res, res ));
@@ -1674,7 +1674,7 @@ static Obj PowObjInt(Obj op, Obj n)
   /* <res> = 0 means that <res> is the neutral element                   */
   else if ( IS_INTOBJ(n) && INT_INTOBJ(n) >   0 ) {
     res = 0;
-    k = 1L << NR_SMALL_INT_BITS;
+    k = (Int)1 << NR_SMALL_INT_BITS;
     l = INT_INTOBJ(n);
     while ( 0 < k ) {
       res = (res == 0 ? res : PROD( res, res ));
